+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- [AWS Lambda]: Instrument deprecated legacy Lambda API (context.done, context.succeed, and context.fail).
+
 ## 1.83.0
 - Add kafkajs instrumentation ([kafkajs](https://www.npmjs.com/package/kafkajs)).